code update and formate

+++ routes/api.php
@@ -126,7 +126,7 @@ Route::prefix('v1')->group(function () {
 
                     Route::get('/', 'index')->middleware(['checkRoleAndPermission:admin,view_service']);
                     Route::get('/{services}', 'show')->middleware(['checkRoleAndPermission:admin,view_service']);
-                    Route::get('/{services}/restore', 'restore');
+                    Route::get('/{services}/restore', 'restore')->middleware(['checkRoleAndPermission:admin,restore_data']);
                     Route::post('/create', 'create')->middleware(['checkRoleAndPermission:admin,create_service']);
                     Route::post('/{service}', 'update')->middleware(['checkRoleAndPermission:admin,edit_service']);
                     Route::delete('/{service}', 'destroy')->middleware(['checkRoleAndPermission:admin,delete_service']);
@@ -138,12 +138,12 @@ Route::prefix('v1')->group(function () {
 
                 Route::controller(ProductController::class)->group(function () {
 
-                    Route::get('/', 'index');
-                    Route::get('/{product}', 'show');
-                    Route::get('/{product}/restore', 'restore');
-                    Route::post('/create', 'create');
-                    Route::post('/{product}', 'update');
-                    Route::delete('/{product}', 'destroy');
+                    Route::get('/', 'index')->middleware(['checkRoleAndPermission:admin,view_product']);
+                    Route::get('/{product}', 'show')->middleware(['checkRoleAndPermission:admin,view_product']);
+                    Route::get('/{product}/restore', 'restore')->middleware(['checkRoleAndPermission:admin,restore_data']);
+                    Route::post('/create', 'create')->middleware(['checkRoleAndPermission:admin,create_product']);
+                    Route::post('/{product}', 'update')->middleware(['checkRoleAndPermission:admin,edit_product']);
+                    Route::delete('/{product}', 'destroy')->middleware(['checkRoleAndPermission:admin,delete_product']);
                 });
             });
 
@@ -152,12 +152,12 @@ Route::prefix('v1')->group(function () {
 
                 Route::controller(ProductCatrgoryController::class)->group(function () {
 
-                    Route::get('/', 'index');
-                    Route::get('/{productcategories}', 'show');
-                    Route::get('/{productcategories}/restore', 'restore');
-                    Route::post('/create', 'create');
-                    Route::post('/{productcategories}', 'update');
-                    Route::delete('/{productcategories}', 'destroy');
+                    Route::get('/', 'index')->middleware(['checkRoleAndPermission:admin,view_productCategory']);
+                    Route::get('/{productcategories}', 'show')->middleware(['checkRoleAndPermission:admin,view_productCategory']);
+                    Route::get('/{productcategories}/restore', 'restore')->middleware(['checkRoleAndPermission:admin,restore_data']);
+                    Route::post('/create', 'create')->middleware(['checkRoleAndPermission:admin,create_productCategory']);
+                    Route::post('/{productcategories}', 'update')->middleware(['checkRoleAndPermission:admin,edit_ProductCategory']);
+                    Route::delete('/{productcategories}', 'destroy')->middleware(['checkRoleAndPermission:admin,delete_productCategory']);
                 });
             });
         });
